Fix compiler errors reported by the Xcode 14 compiler.
Remove `supp_size` variable from hb-subset-cff1.cc.  The variable was set
but not used causing a compiler warning that was treated as an error.
This change already exists in the upstream chromium repo.

Fix a couple of ambiguous conversions from FloatRect to gfx::RectF().
The change in TextFinder::ActiveFindMatchRect() already exists
in the upstream chromium repo.  The change in TextFinder::FindMatchRects()
is a suitable workaround.  I could not apply the change from upstream
because it depended on other significant changes.